src/backend/requirements.txt
fastapi
uvicorn
torch
requests
numpy
transformers
pillow
python-multipart